requirements.txt
crewai>=0.28.0
crewai-tools>=0.12.0
python-dotenv>=1.0.0
langchain-groq>=0.1.0